requirements.txt
tensorflow
numpy
pyyaml
tokenizers
